requirements.txt
asn1crypto==0.22.0
astroid==1.5.3
backports.functools-lru-cache==1.4
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
configparser==3.5.0
Django==1.7
django-facebook==6.0.3
dnspython==1.15.0
enum34==1.1.6
idna==2.6
ipaddress==1.0.18
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
MySQL-python==1.2.5
Naked==0.1.31
olefile==0.44
Pillow==4.2.1
pycparser==2.18
pycrypto==2.6.1
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
PyYAML==3.12
requests==2.18.4
shellescape==3.4.1
singledispatch==3.4.0.3
six==1.10.0
Unidecode==0.4.21
urllib3==1.22
wrapt==1.10.11
